Update index.html
Major changes:
Added section: "Your Units Being Countered"
Added section: "Units to Avoid Purchasing"

Your units being countered:  These are your current units that the enemy's current units are counters for.

Units to avoid purchasing:  These are units that are weak to units the enemy currently has out.

Where is this data from:  This data is in the countersData array.  That array contains 2 arrays.  The first, "weakness", is what this unit is weak toward and the second, "strong_counter", is what this unit is strong against.   Here is an example (data may have changed):
                "Arclight": { "weakness": ["Stormcaller", "Steel Ball", "Rhino", "Scorpion", "Vulcan", "Overlord", "Melting Point", "Fortress", "War Factory"], "strong_counter": ["Crawler", "Hounds", "Steel Ball", "Fang", "Mustang"] },